% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{applications_case_study_leeds}
\alias{applications_case_study_leeds}
\title{Example of planning applications dataset from the PlanIt API}
\description{
Example of planning applications dataset from the PlanIt API
}
\note{
The schema data can be (re-)generated using case studies vignette.
}
\examples{
applications_case_study_leeds
plot(applications_case_study_leeds)
}
\keyword{datasets}
