% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planit.R
\name{bbox_to_string}
\alias{bbox_to_string}
\title{Convert bbox to string}
\usage{
bbox_to_string(bbox)
}
\arguments{
\item{bbox}{Bounding box in the form of xmin, ymin, xmax, ymax, e.g.: `c(-1.3, 53.7, -1.2, 53.9)`}
}
\description{
Convert bbox to string
}
