% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provision.R
\name{estimate_walkability}
\alias{estimate_walkability}
\title{Estimate walkability}
\usage{
estimate_walkability(
  osm_network,
  parameters = c(hilliness = 1.3, directness = 1.8),
  walking_potential_weights = NULL
)
}
\arguments{
\item{osm_network}{sf object of highways from OSM}

\item{parameters}{Weightings given to different columns (some of which could be calculated)}

\item{walking_potential_weights}{Optional variable to weight the measure by estimated number of trips on different segments (by scenario)}
}
\value{

}
\description{
Estimate walkability
}
