% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planit.R
\name{get_planit_url}
\alias{get_planit_url}
\title{Get url of planit API key}
\usage{
get_planit_url(
  bbox,
  base_url = "https://www.planit.org.uk/api/applics/json",
  limit = 30,
  end_date = "2009-02-01",
  start_date = "2000-02-01"
)
}
\arguments{
\item{bbox}{Bounding box in the form of xmin, ymin, xmax, ymax, e.g.: `c(-1.3, 53.7, -1.2, 53.9)`}

\item{base_url}{The base URL of the service}

\item{limit}{How many items to return (e.g. 6, default)}

\item{end_date}{E.g. `"2009-02-01"`}

\item{start_date}{E.g. `"2000-02-01"`}
}
\description{
Get url of planit API key
}
