% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{get_jts_data}
\alias{get_jts_data}
\alias{jts_tables}
\title{Get accessibility data for particular year and table number}
\usage{
get_jts_data(table, year = NULL, u_csv = NULL, skip = 6)
}
\arguments{
\item{table}{The title of the table, e.g. "jts0501"}

\item{year}{The year, e.g. 2017}

\item{u_csv}{The csv url (not usually needed)}

\item{skip}{How many lines to skip before the data starts (6 by default)}
}
\description{
The DfT's Journey Time Statistics are outlined here on the
[gov.uk website](https://www.gov.uk/government/statistical-data-sets/journey-time-statistics-data-tables-jts).
}
\details{
The tables starting JTS01 to JTS03 provide national overview data.
The tables JTS0401 to JTS0409 provide data at the Local Authority level.
The tables JTS0501 to JTS0509 provide the same data at the LSOA level.
The tables beginning JTS09 provide data on accessibility to transport hubs.
And the tables beginning JTS10 contain add other variables.

Data is provided every year from 2014 to 2019 in many cases
}
\examples{
jts_tables
jts_tables$table_title
get_jts_data(table = "jts0401", year = 2017)
}
