% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planit.R
\name{get_planit_data}
\alias{get_planit_data}
\title{Get data from Planit API}
\usage{
get_planit_data(
  query_type = "applics",
  query_value = NULL,
  fmt = "geojson",
  base_url = "https://dev.planit.org.uk",
  limit = 6,
  end_date = as.character(Sys.Date()),
  start_date = "2000-02-01",
  auth = NULL,
  pcode = NULL,
  krad = NULL,
  bbox = NULL,
  app_size = NULL,
  app_state = NULL,
  app_type = NULL,
  silent = FALSE
)
}
\arguments{
\item{query_type}{The type of query (`"applics"` by default)}

\item{query_value}{Text string associated with the `query_type`.
To be used only when `query_type` is not `"applics"`.}

\item{fmt}{The format of the output (`"geojson"` returns an `sf` object, `"json"` returns a data frame)}

\item{base_url}{The base URL of the service}

\item{limit}{How many items to return (e.g. 6, default)}

\item{end_date}{E.g. `"2009-02-01"`. Default is `as.character(Sys.Date())`.}

\item{start_date}{The earliest application (date of application) to be filtered `"2000-02-01"`}

\item{auth}{Authority name}

\item{pcode}{Postcode = UK postcode to use for the centre of a location search}

\item{krad}{Radius (km) = only planning applications within the circle perimeter are returned (default 2)}

\item{bbox}{Bounding box in the form of xmin, ymin, xmax, ymax, e.g.: `c(-1.3, 53.7, -1.2, 53.9)`}

\item{app_size}{PlanIt classification for the size of planning applications. Values are `"large"` `"medium"` and `"small"`}

\item{app_state}{Current decision status of the planning application (Undecided, Permitted, Conditions, Rejected, Withdrawn, Referred, Unresolved, Other)}

\item{app_type}{Type of the application (Full, Outline, Amendment, Heritage, Trees, Advertising, Telecoms, Other)}

\item{silent}{Do you want a message? Default is `FALSE`}
}
\value{
A (geographic) data frame
}
\description{
This function requests data from [planit.org.uk](https://www.planit.org.uk).
See their [API spec](https://www.planit.org.uk/api/) and their
[data dictionary](https://www.planit.org.uk/dictionary/)
for further information.
}
\examples{
# tmaptools::bb("university leeds") # find a bounding box
bbox = c(-1.5, 53.7, -1.3, 53.8)
res = get_planit_data(bbox = bbox, silent = FALSE) # return geographic (`sf`) object
class(res)
names(res)
plot(sf::st_geometry(res))
get_planit_data(fmt = "json", limit = 2, bbox = bbox) # return data frame with limit
get_planit_data(fmt = "json", limit = 2, auth = "leeds")$description
get_planit_data(end_date = "2008-01-01", limit = 2, bbox = bbox)$description # historic data
get_planit_data(pcode = "LS2 9JT", limit = 2, bbox = bbox)$description # data from specific postcode
get_planit_data(query_type = "planapplic", query_value = "13/05235/FU@Leeds")$description
get_planit_data(query_type = "planarea", query_value = "leeds")$description
leeds_area = get_planit_data(query_type = "areas", limit = 2, auth = "Leeds")
plot(leeds_area) # geographic outline of the area
planitareas = get_planit_data(query_type = "areas", limit = 2)
nrow(planitareas)
planitareas$name
planitareas
}
