% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planit.R
\name{get_planit_data}
\alias{get_planit_data}
\title{Get data from Planit API}
\usage{
get_planit_data(
  bbox,
  query_type = "applics",
  query_type_search = NULL,
  fmt = "geojson",
  base_url = "https://www.planit.org.uk/api",
  limit = 6,
  end_date = as.character(Sys.Date()),
  start_date = "2000-02-01",
  pcode = NULL,
  krad = NULL,
  silent = FALSE
)
}
\arguments{
\item{bbox}{Bounding box in the form of xmin, ymin, xmax, ymax, e.g.: `c(-1.3, 53.7, -1.2, 53.9)`}

\item{query_type}{The type of query (`"applics"` by default)}

\item{query_type_search}{Text string associated with the `query_type` (may be updated)}

\item{fmt}{The format of the output (`"geojson"` returns an `sf` object, `"json"` returns a data frame)}

\item{base_url}{The base URL of the service}

\item{limit}{How many items to return (e.g. 6, default)}

\item{end_date}{E.g. `"2009-02-01"`}

\item{start_date}{E.g. `"2000-02-01"`}

\item{pcode}{Postcode = UK postcode to use for the centre of a location search}

\item{krad}{Radius (km) = only planning applications within the circle perimeter are returned (default 2)}

\item{silent}{Do you want a message? Default is `FALSE`}
}
\value{
A (geographic) data frame
}
\description{
This function requests data from [planit.org.uk](https://www.planit.org.uk).
See their [API spec](https://www.planit.org.uk/api/) and their
[data dictionary](https://www.planit.org.uk/dictionary/)
for further information.
}
\examples{
bbox = c(-1.366023, 53.744171, -1.35515, 53.747495)
get_planit_data(bbox) # return geographic (`sf`) object
get_planit_data(bbox, fmt = "json") # return data frame
get_planit_data(bbox, end_date = "2008-01-01") # historic data
get_planit_data(bbox, pcode = "LS2 9JT") # data from specific postcode
}
